added CRUD operations, configured cache and read-only transactions

+++ src/main/java/com/api/FastShopping/products/dtos/ProductDTO.java
@@ -1,7 +1,9 @@
 package com.api.FastShopping.products.dtos;
 
+import java.math.BigDecimal;
+
 public record ProductDTO(
         String name,
-        Float price
+        BigDecimal price
 ) {
 }
